Reference Problem
This commit does not compile.  GCC is throwing an error I've never seen that is preventing me from building the factory function idiomatically, as we discussed in your office.  The only way I've been able to get this to work when ReadData() is a friend function (instead of member function) of CSVParser is if I remove both references (return type and argument).  In which case there is most likely a copy operation on the massive set of data.  I'd like to leave the factory function as I had it earlier today, returning *this and avoiding all copy operations on CSVParser.